:bug: FIX: fix integer type edge cases
- Add Value::Int case to for loop iteration (was only handling Float)
- Add Value::Int case to IndexAccess for arrays and strings
- Change current_timestamp() to return Int instead of Float (timestamps are integers)
- Fix error.line property to return Int instead of Float
- Fix all test assertions to match new Int/Float types
- All 127 tests now passing

+++ src/builtins.rs
@@ -504,11 +504,11 @@ pub fn now() -> f64 {
 /// Get current timestamp in milliseconds since UNIX epoch
 /// Returns the number of milliseconds elapsed since January 1, 1970 00:00:00 UTC
 /// This is useful for timestamps and timing operations
-pub fn current_timestamp() -> f64 {
+pub fn current_timestamp() -> i64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .expect("System time before UNIX epoch")
-        .as_millis() as f64
+        .as_millis() as i64
 }
 
 /// High-resolution performance timer in milliseconds
